# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OsanVault Africa API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Property {
  id: number;
  name: string;
  type: string;
  location: string;
  country: string;
  flag: string;
  targetRaise: number;
  raised: number;
  totalTokens: number;
  tokenPrice: number;
  yieldApy: number;
  size: string;
  status: string;
  carbonOffsetTonnes: number;
  description: string;
  gradientFrom: string;
  gradientTo: string;
  jurisdiction: string;
  indigenousAuthority: string;
  legalDocCid: string;
  lat: number;
  lng: number;
  createdAt: string;
  updatedAt: string;
}

export interface CarbonProject {
  id: number;
  name: string;
  methodology: string;
  region: string;
  vintage: number;
  totalIssued: string;
  totalRetired: string;
  verified: boolean;
  verifier: string;
  flag: string;
  description: string;
  lat: number;
  lng: number;
  registryLink: string;
  linkedPropertyId?: number | null;
  createdAt: string;
}

export interface GovernanceProposal {
  id: number;
  title: string;
  description: string;
  proposer: string;
  status: string;
  votesFor: number;
  votesAgainst: number;
  quorum: number;
  endTime: string;
  category: string;
  createdAt: string;
}

export interface ActivityEvent {
  id: number;
  type: string;
  propertyId?: number | null;
  propertyName?: string | null;
  projectId?: number | null;
  projectName?: string | null;
  amount: string;
  amountNgn?: number | null;
  address: string;
  txHash: string;
  timestamp: string;
  createdAt: string;
}

export interface PlatformStats {
  propertiesLive: number;
  tvlNgn: number;
  totalInvestors: number;
  osanvStaked: number;
  totalCarbonTonnes: number;
  avgPropertyYield: number;
}

export type ListActivityParams = {
limit?: number;
};