# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: OsanVault Africa API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a single property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Property:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        location:
          type: string
        country:
          type: string
        flag:
          type: string
        targetRaise:
          type: number
        raised:
          type: number
        totalTokens:
          type: integer
        tokenPrice:
          type: integer
        yieldApy:
          type: number
        size:
          type: string
        status:
          type: string
        carbonOffsetTonnes:
          type: integer
        description:
          type: string
        gradientFrom:
          type: string
        gradientTo:
          type: string
        jurisdiction:
          type: string
        indigenousAuthority:
          type: string
        legalDocCid:
          type: string
        lat:
          type: number
        lng:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - location
        - country
        - flag
        - targetRaise
        - raised
        - totalTokens
        - tokenPrice
        - yieldApy
        - size
        - status
        - carbonOffsetTonnes
        - description
        - gradientFrom
        - gradientTo
        - jurisdiction
        - indigenousAuthority
        - legalDocCid
        - lat
        - lng
        - createdAt
        - updatedAt
...