# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OsanVault Africa API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Property {
  id: number;
  name: string;
  type: string;
  location: string;
  country: string;
  flag: string;
  targetRaise: number;
  raised: number;
  totalTokens: number;
  tokenPrice: number;
  yieldApy: number;
  size: string;
  status: string;
  carbonOffsetTonnes: number;
  description: string;
  gradientFrom: string;
  gradientTo: string;
  jurisdiction: string;
  indigenousAuthority: string;
  legalDocCid: string;
  lat: number;
  lng: number;
  createdAt: Date;
  updatedAt: Date;
}